pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "flakeradar"
version = "0.1.0"
description = "Local flakiness analyzer with AI summaries + Redis/Kafka stubs."
authors = [{name="Harsh", email="you@example.com"}]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "click>=8.1",
  "lxml>=5.2",
  "tabulate>=0.9",
  "jinja2>=3.1",
  "requests>=2.32",
  "openai>=1.40",           # official client
  "redis>=5.0",
  "kafka-python>=2.0",      # lightweight; swap for confluent-kafka later
  "sqlalchemy>=2.0",
  "alembic>=1.13",
  "pydantic>=2.7"
]

[project.scripts]
flakeradar = "flakeradar.cli:main"